+++ src/lib.rs
@@ -27,6 +27,16 @@
 //! }
 //! ```
 
+mod builder;
+mod connector;
+mod error;
+mod request_kind;
+mod result;
+
 /// Tools for communicating with the server using sessions.
+pub mod insecure_connector;
+pub mod secure_connector;
 pub mod session;
+pub use insecure_connector::InsecureConnector;
+pub use secure_connector::SecureConnector;
 pub use session::Session;
